# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Writing Project API
  description: |
    記事作成支援Webアプリ「おひとりさまライフ」のAPI仕様書。

    ## 概要

    本APIは2つの用途に分かれています:

    1. **管理用API** (`/api/articles`): 管理画面から記事のCRUD操作を行うためのAPI。NextAuthのセッション認証が必要です。
    2. **外部公開用API** (`/api/publish`): 外部システム（AIエージェント等）から記事を投稿するためのAPI。APIキーによるBearer認証が必要です。
    3. **OGP画像生成API** (`/api/og`): 記事のOGP画像を動的に生成するAPI。認証不要です。

    ## 認証方式

    - **セッション認証 (sessionAuth)**: NextAuth v5によるGoogle OAuth + JWTセッション。ホワイトリストに登録されたメールアドレスのみログイン可能。管理画面の `/articles/*` ページおよび管理用APIで使用。
    - **APIキー認証 (apiKeyAuth)**: `Authorization: Bearer <API_KEY>` ヘッダーによる認証。外部システムからの記事投稿で使用。
  version: "1.0.0"
  contact:
    name: Writing Project
  license:
    name: Private
    url: https://example.com

servers:
  - url: http://localhost:3000
    description: ローカル開発環境
  - url: https://production.example.com
    description: 本番環境

tags:
  - name: articles
    description: |
      管理用記事API。管理画面（`/articles`）から使用されるCRUDエンドポイント。
      NextAuthセッション認証が必要です（ミドルウェアにより `/articles/*` ページアクセス時に認証チェックが行われます）。
  - name: publish
    description: |
      外部公開用記事投稿API。AIエージェントや外部システムから記事を投稿するためのエンドポイント。
      APIキーによるBearer認証が必要です。
  - name: og
    description: |
      OGP画像動的生成API。記事のタイトルとカテゴリからOGP画像を生成します。
      認証不要。Edge Runtimeで動作します。
  - name: auth
    description: |
      NextAuth認証エンドポイント。Google OAuth 2.0によるログイン・ログアウト処理。
      NextAuthが自動的にハンドリングするため、直接呼び出す必要はありません。

paths:
  /api/articles:
    get:
      tags:
        - articles
      summary: 管理用記事一覧取得
      description: |
        管理画面で使用する記事一覧を取得します。

        **利用者**: 管理画面UI（認証済みユーザー）

        ステータス、ソート順、キーワード検索でフィルタリングが可能です。
        デフォルトでは作成日時の降順（新しい順）で全記事が返されます。

        検索クエリは記事のタイトル、本文、キーワードに対して部分一致検索を行います。

        **注意**: 現在のAPIルート実装にはセッション認証チェックが含まれていませんが、
        ミドルウェアにより `/articles/*` ページからのアクセス時に認証が強制されます。
        今後、APIルートレベルでの認証チェック追加が推奨されます。
      operationId: listArticles
      parameters:
        - name: status
          in: query
          required: false
          description: |
            記事のステータスでフィルタリングします。
            指定しない場合は全ステータスの記事が返されます。
          schema:
            $ref: "#/components/schemas/ArticleStatus"
        - name: sortBy
          in: query
          required: false
          description: |
            ソートの基準となるフィールドを指定します。
            デフォルトは `createdAt`（作成日時）です。
          schema:
            type: string
            enum:
              - createdAt
              - updatedAt
              - title
            default: createdAt
        - name: sortOrder
          in: query
          required: false
          description: |
            ソート順を指定します。
            デフォルトは `desc`（降順）です。
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
        - name: searchQuery
          in: query
          required: false
          description: |
            検索キーワード。記事のタイトル、本文、キーワードに対して
            大文字小文字を区別しない部分一致検索を行います。
          schema:
            type: string
          example: Next.js
      responses:
        "200":
          description: 記事一覧の取得に成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
              example:
                - id: "550e8400-e29b-41d4-a716-446655440000"
                  title: "Next.js 15の新機能まとめ"
                  content: "# Next.js 15\n\nNext.js 15では..."
                  keywords:
                    - Next.js
                    - React
                    - フロントエンド
                  status: published
                  createdAt: "2026-03-15T10:30:00.000Z"
                  updatedAt: "2026-03-20T14:00:00.000Z"
                  publishedAt: "2026-03-18T09:00:00.000Z"
                - id: "550e8400-e29b-41d4-a716-446655440001"
                  title: "資産形成の基礎"
                  content: ""
                  keywords: []
                  status: draft
                  createdAt: "2026-03-21T08:00:00.000Z"
                  updatedAt: "2026-03-21T08:00:00.000Z"
                  publishedAt: null
        "500":
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags:
        - articles
      summary: 管理用記事作成
      description: |
        新しい記事を作成します。

        **利用者**: 管理画面UI（認証済みユーザー）

        **認証**: NextAuthセッション認証が必要です。

        タイトルは必須です。本文、キーワード、ステータスは省略可能で、
        省略した場合はそれぞれ空文字列、空配列、`draft` がデフォルト値として設定されます。

        記事IDはUUID v4で自動生成されます。
        `createdAt` と `updatedAt` は現在時刻で自動設定されます。
        `publishedAt` は初期値 `null` で、ステータスが `published` に変更された時点で設定されます。
      operationId: createArticle
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleCreateInput"
            examples:
              minimal:
                summary: 最小限の記事作成（タイトルのみ）
                value:
                  title: "新しい記事"
              full:
                summary: 全フィールド指定
                value:
                  title: "TypeScriptの型システム入門"
                  content: "# TypeScriptの型システム\n\nTypeScriptは..."
                  keywords:
                    - TypeScript
                    - 型システム
                    - プログラミング
                  status: draft
      responses:
        "201":
          description: 記事の作成に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
              example:
                id: "550e8400-e29b-41d4-a716-446655440002"
                title: "TypeScriptの型システム入門"
                content: "# TypeScriptの型システム\n\nTypeScriptは..."
                keywords:
                  - TypeScript
                  - 型システム
                  - プログラミング
                status: draft
                createdAt: "2026-03-21T10:00:00.000Z"
                updatedAt: "2026-03-21T10:00:00.000Z"
                publishedAt: null
        "400":
          description: |
            リクエストのバリデーションエラー。以下のケースで発生します:
            - `title` が未指定または文字列でない
            - `content` が文字列でない（指定した場合）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                titleMissing:
                  summary: タイトル未指定
                  value:
                    error: "Title is required"
                contentInvalid:
                  summary: 本文が文字列でない
                  value:
                    error: "Content must be a string"
        "401":
          description: 認証エラー。セッションが無効または期限切れです。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/articles/{id}:
    get:
      tags:
        - articles
      summary: 管理用記事詳細取得
      description: |
        指定されたIDの記事詳細を取得します。

        **利用者**: 管理画面UI（認証済みユーザー）

        記事が見つからない場合は404エラーを返します。
      operationId: getArticle
      parameters:
        - $ref: "#/components/parameters/ArticleId"
      responses:
        "200":
          description: 記事詳細の取得に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                title: "Next.js 15の新機能まとめ"
                content: "# Next.js 15\n\nNext.js 15では..."
                keywords:
                  - Next.js
                  - React
                  - フロントエンド
                status: published
                createdAt: "2026-03-15T10:30:00.000Z"
                updatedAt: "2026-03-20T14:00:00.000Z"
                publishedAt: "2026-03-18T09:00:00.000Z"
        "404":
          description: 指定されたIDの記事が見つかりません
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Article not found"
        "500":
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags:
        - articles
      summary: 管理用記事更新
      description: |
        指定されたIDの記事を更新します。

        **利用者**: 管理画面UI（認証済みユーザー）

        **認証**: NextAuthセッション認証が必要です。

        リクエストボディに含まれたフィールドのみ更新されます（部分更新）。
        `updatedAt` は自動的に現在時刻に更新されます。

        ステータスが `published` に変更された場合（かつ以前のステータスが `published` でない場合）、
        `publishedAt` が現在時刻に自動設定されます。
        既に `published` だった記事の `publishedAt` は変更されません。
      operationId: updateArticle
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/ArticleId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleUpdateInput"
            examples:
              titleOnly:
                summary: タイトルのみ更新
                value:
                  title: "更新されたタイトル"
              publish:
                summary: 記事を公開
                value:
                  status: published
              fullUpdate:
                summary: 全フィールド更新
                value:
                  title: "更新されたタイトル"
                  content: "# 更新された本文\n\n新しい内容..."
                  keywords:
                    - 更新
                    - テスト
                  status: published
      responses:
        "200":
          description: 記事の更新に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                title: "更新されたタイトル"
                content: "# 更新された本文\n\n新しい内容..."
                keywords:
                  - 更新
                  - テスト
                status: published
                createdAt: "2026-03-15T10:30:00.000Z"
                updatedAt: "2026-03-21T12:00:00.000Z"
                publishedAt: "2026-03-21T12:00:00.000Z"
        "401":
          description: 認証エラー。セッションが無効または期限切れです。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 指定されたIDの記事が見つかりません
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Article not found"
        "500":
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - articles
      summary: 管理用記事削除
      description: |
        指定されたIDの記事を削除します。

        **利用者**: 管理画面UI（認証済みユーザー）

        **認証**: NextAuthセッション認証が必要です。

        削除は物理削除（完全に削除）です。復元はできません。
        記事が見つからない場合は404エラーを返します。
      operationId: deleteArticle
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/ArticleId"
      responses:
        "200":
          description: 記事の削除に成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                properties:
                  success:
                    type: boolean
                    example: true
              example:
                success: true
        "401":
          description: 認証エラー。セッションが無効または期限切れです。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 指定されたIDの記事が見つかりません
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Article not found"
        "500":
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/publish:
    post:
      tags:
        - publish
      summary: 外部公開用記事投稿
      description: |
        外部システムから記事を投稿するためのエンドポイントです。

        **利用者**: AIエージェント、外部システム（管理画面からは使用しません）

        **認証**: APIキーによるBearer認証が必要です。
        `Authorization: Bearer <PUBLISH_API_KEY>` ヘッダーを送信してください。

        投稿された記事はMDXファイルとしてコンテンツディレクトリに保存されます。
        管理用APIの記事とは異なるストレージに保存される点にご注意ください。

        **レート制限に関する注意**: 現在レート制限は実装されていませんが、
        本番環境では適切なレート制限の導入が推奨されます。

        **カテゴリ**: `asset`（資産形成）、`tech`（プログラミング）、`health`（健康）のいずれかを指定してください。
      operationId: publishArticle
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishArticleInput"
            example:
              title: "資産形成の基礎知識"
              content: "# 資産形成の基礎\n\n資産形成とは..."
              category: asset
              slug: "asset-basics-2026"
              tags:
                - 資産形成
                - 投資
                - 初心者向け
              description: "資産形成の基本的な考え方と具体的な方法を解説します。"
              thumbnail: "/images/asset-basics.png"
      responses:
        "201":
          description: 記事の投稿に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishArticleResponse"
              example:
                success: true
                slug: "asset-basics-2026"
                url: "/asset/asset-basics-2026"
        "400":
          description: |
            リクエストのバリデーションエラー。必須フィールドの不足や不正な値が含まれる場合に発生します。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "title, content, category, and slug are required"
        "401":
          description: |
            認証エラー。APIキーが未指定、不正、または期限切れの場合に発生します。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Invalid or missing API key"
        "409":
          description: |
            スラッグの競合。指定されたスラッグの記事が既に存在する場合に発生します。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Article with this slug already exists"
        "500":
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/og:
    get:
      tags:
        - og
      summary: OGP画像動的生成
      description: |
        記事のタイトルとカテゴリからOGP（Open Graph Protocol）画像を動的に生成します。

        **利用者**: SNSクローラー、ブラウザ（metaタグから自動的にアクセスされます）

        **認証**: 不要

        **ランタイム**: Edge Runtime（Vercel Edge Functions）で動作します。

        生成される画像は1200x630ピクセルのPNG形式です。
        カテゴリに応じてテーマカラーが変わります:
        - `asset`: 淡い紫背景 + 青アクセント（ラベル: 資産形成）
        - `tech`: ダーク背景 + 青アクセント（ラベル: プログラミング）
        - `health`: 白背景 + 青アクセント（ラベル: 健康）
        - 未指定/その他: 白背景 + 青アクセント（ラベルなし）

        タイトルは40文字を超える場合、末尾が省略記号（...）で切り詰められます。

        **レート制限に関する注意**: Edge Runtimeで動作するため、Vercelのエッジ関数の制限に従います。
        CDNキャッシュの設定が推奨されます。
      operationId: generateOgImage
      parameters:
        - name: title
          in: query
          required: false
          description: |
            OGP画像に表示するタイトル。
            省略した場合は「おひとりさまライフ」がデフォルト値として使用されます。
            40文字を超える場合は自動的に切り詰められます。
          schema:
            type: string
            default: "おひとりさまライフ"
          example: "Next.js 15の新機能まとめ"
        - name: category
          in: query
          required: false
          description: |
            記事のカテゴリ。画像のテーマカラーとカテゴリラベルに影響します。
            省略した場合はデフォルトテーマ（白背景、ラベルなし）が使用されます。
          schema:
            $ref: "#/components/schemas/Category"
          example: tech
      responses:
        "200":
          description: OGP画像の生成に成功
          content:
            image/png:
              schema:
                type: string
                format: binary
          headers:
            Content-Type:
              schema:
                type: string
                example: image/png
        "500":
          description: 画像生成エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/auth/{nextauth}:
    get:
      tags:
        - auth
      summary: NextAuth認証エンドポイント（GET）
      description: |
        NextAuth v5が自動的にハンドリングする認証関連エンドポイントです。
        直接呼び出すことは通常ありません。

        **利用者**: NextAuthクライアント（自動）

        以下のパスがサポートされています:
        - `/api/auth/signin` - サインインページ
        - `/api/auth/signout` - サインアウト処理
        - `/api/auth/callback/google` - Google OAuthコールバック
        - `/api/auth/session` - セッション情報取得
        - `/api/auth/csrf` - CSRFトークン取得
        - `/api/auth/providers` - 利用可能なプロバイダー一覧

        **認証方式**: Google OAuth 2.0。ホワイトリストに登録されたメールアドレスのみログインが許可されます。
        セッションはJWT戦略で管理されます。

        未認証ユーザーが `/articles/*` にアクセスすると、ミドルウェアにより `/login` にリダイレクトされます。
      operationId: nextAuthGet
      parameters:
        - name: nextauth
          in: path
          required: true
          description: NextAuthの内部ルーティングパス（catchall）
          schema:
            type: string
          example: "session"
      responses:
        "200":
          description: リクエストに応じたレスポンス（セッション情報、HTMLページ等）
        "302":
          description: OAuth認証フローでのリダイレクト
    post:
      tags:
        - auth
      summary: NextAuth認証エンドポイント（POST）
      description: |
        NextAuth v5が自動的にハンドリングするPOST系認証エンドポイントです。
        サインイン・サインアウトのフォーム送信処理等で使用されます。
        直接呼び出すことは通常ありません。
      operationId: nextAuthPost
      parameters:
        - name: nextauth
          in: path
          required: true
          description: NextAuthの内部ルーティングパス（catchall）
          schema:
            type: string
          example: "signin/google"
      responses:
        "200":
          description: 処理成功
        "302":
          description: 認証フロー完了後のリダイレクト

components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: authjs.session-token
      description: |
        NextAuth v5によるCookieベースのセッション認証。

        Google OAuth 2.0でログイン後、JWTセッショントークンがCookieに保存されます。
        管理画面（`/articles/*`）へのアクセスおよび管理用API（POST/PUT/DELETE）で使用されます。

        ホワイトリストに登録されたメールアドレスのみログインが許可されます。
        セッションの有効期限はNextAuthのデフォルト設定（30日）に従います。

    apiKeyAuth:
      type: http
      scheme: bearer
      description: |
        APIキーによるBearer認証。

        `Authorization: Bearer <PUBLISH_API_KEY>` ヘッダーを使用します。
        外部システム（AIエージェント等）からの記事投稿API（`/api/publish`）で使用されます。

        APIキーは環境変数 `PUBLISH_API_KEY` で設定します。

  parameters:
    ArticleId:
      name: id
      in: path
      required: true
      description: |
        記事のユニークID。UUID v4形式です。
        `crypto.randomUUID()` により自動生成されます。
      schema:
        type: string
        format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"

  schemas:
    ArticleStatus:
      type: string
      enum:
        - draft
        - published
        - archived
      description: |
        記事のステータス。
        - `draft`: 下書き。作成直後のデフォルト状態。
        - `published`: 公開済み。ステータスをpublishedに変更すると `publishedAt` が自動設定されます。
        - `archived`: アーカイブ済み。非公開だが削除はされていない状態。

    Category:
      type: string
      enum:
        - asset
        - tech
        - health
      description: |
        記事カテゴリ。
        - `asset`: 資産形成
        - `tech`: プログラミング
        - `health`: 健康

    Article:
      type: object
      required:
        - id
        - title
        - content
        - keywords
        - status
        - createdAt
        - updatedAt
        - publishedAt
      properties:
        id:
          type: string
          format: uuid
          description: 記事のユニークID（UUID v4）
          example: "550e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          description: 記事のタイトル
          example: "Next.js 15の新機能まとめ"
        content:
          type: string
          description: 記事の本文（Markdown形式）
          example: "# Next.js 15\n\nNext.js 15では..."
        keywords:
          type: array
          items:
            type: string
          description: 記事に関連するキーワード一覧
          example:
            - Next.js
            - React
            - フロントエンド
        status:
          $ref: "#/components/schemas/ArticleStatus"
        createdAt:
          type: string
          format: date-time
          description: 記事の作成日時（ISO 8601形式）
          example: "2026-03-15T10:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          description: 記事の最終更新日時（ISO 8601形式）
          example: "2026-03-20T14:00:00.000Z"
        publishedAt:
          type: string
          format: date-time
          nullable: true
          description: |
            記事の公開日時（ISO 8601形式）。
            ステータスが `published` に変更された時点で自動設定されます。
            未公開の場合は `null` です。
          example: "2026-03-18T09:00:00.000Z"

    ArticleCreateInput:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: 記事のタイトル（必須）
          example: "TypeScriptの型システム入門"
        content:
          type: string
          description: 記事の本文（Markdown形式）。省略時は空文字列。
          default: ""
          example: "# TypeScriptの型システム\n\nTypeScriptは..."
        keywords:
          type: array
          items:
            type: string
          description: 記事に関連するキーワード一覧。省略時は空配列。
          default: []
          example:
            - TypeScript
            - 型システム
        status:
          $ref: "#/components/schemas/ArticleStatus"
          description: 記事のステータス。省略時は `draft`。
          default: draft

    ArticleUpdateInput:
      type: object
      description: |
        記事の更新入力。すべてのフィールドが省略可能です。
        指定されたフィールドのみ更新されます（部分更新）。
      properties:
        title:
          type: string
          description: 記事のタイトル
          example: "更新されたタイトル"
        content:
          type: string
          description: 記事の本文（Markdown形式）
          example: "# 更新された本文"
        keywords:
          type: array
          items:
            type: string
          description: 記事に関連するキーワード一覧
          example:
            - 更新
            - テスト
        status:
          $ref: "#/components/schemas/ArticleStatus"
          description: 記事のステータス

    PublishArticleInput:
      type: object
      required:
        - title
        - content
        - category
        - slug
      properties:
        title:
          type: string
          description: 記事のタイトル（必須）
          example: "資産形成の基礎知識"
        content:
          type: string
          description: 記事の本文（MDX形式）（必須）
          example: "# 資産形成の基礎\n\n資産形成とは..."
        category:
          $ref: "#/components/schemas/Category"
          description: 記事のカテゴリ（必須）
        slug:
          type: string
          description: |
            記事のURL用スラッグ（必須）。
            URLの一部として使用されるため、半角英数字とハイフンのみ推奨。
            既存の記事と重複する場合は409エラーが返されます。
          pattern: "^[a-z0-9]+(-[a-z0-9]+)*$"
          example: "asset-basics-2026"
        tags:
          type: array
          items:
            type: string
          description: 記事のタグ一覧（省略可能）
          example:
            - 資産形成
            - 投資
            - 初心者向け
        description:
          type: string
          description: 記事の概要説明。metaタグのdescriptionに使用されます。（省略可能）
          example: "資産形成の基本的な考え方と具体的な方法を解説します。"
        thumbnail:
          type: string
          description: サムネイル画像のパス（省略可能）
          example: "/images/asset-basics.png"

    PublishArticleResponse:
      type: object
      required:
        - success
        - slug
        - url
      properties:
        success:
          type: boolean
          description: 投稿成功フラグ
          example: true
        slug:
          type: string
          description: 投稿された記事のスラッグ
          example: "asset-basics-2026"
        url:
          type: string
          description: 投稿された記事のURL
          example: "/asset/asset-basics-2026"

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: エラーメッセージ
          example: "An internal error occurred"